refactor contact and home section by separating logic for opening modal to smaller components
