+++ src/components/sections/contact/Contact.module.css
@@ -1,12 +1,13 @@
 .container {
     display: flex;
     column-gap: var(--spacing-XS);
-    .icon {
-        color: var(--background-color-primary);
-        &:hover {
-            transform: scale(1.2);
-            color: aqua; /* Change the color of the icon */
-        }
+}
+
+.icon {
+    color: var(--background-color-primary);
+    &:hover {
+        transform: scale(1.2);
+        color: aqua; /* Change the color of the icon */
     }
 }
